update analytics override for hyrax-3.6
